AlarmC and CounterC now support both 32khz and Milli.
It might be nice to split up the 32khz and Milli components because, for
instance, if the 32khz counter is never used, then it'd be nice to generate
code that doesn't include it.  Currently it will be included even if it
isn't used because it's handling interrupts to manage the counter itself, and
GCC can't exactly know that that code isn't necessary.  I think there'd be a
nice solution with Impl's, but I'd have to think about it for a moment to
make sure it all works out okay.

Also, you'll note that I'm wiring in MathOpsM and CastOpsM explicitly here
instead of via say TransformAlarmC and TranformCounterC in system.  That's
because the nesC wiring deductions are working -- I need to show David G.